+++ tests/unit/test_doubles_test.sh
@@ -34,7 +34,7 @@ function test_successful_spy() {
   ps a_random_parameter_1 a_random_parameter_2
 
   assert_have_been_called_with "a_random_parameter_1 a_random_parameter_2" ps
-  assertHaveBeenCalled ps
+  assert_have_been_called ps
 }
 
 function test_unsuccessful_spy_called() {
@@ -42,7 +42,7 @@ function test_unsuccessful_spy_called() {
 
   assert_equals\
     "$(Console::printFailedTest "Unsuccessful spy called" "ps" "has not been called at least" "once")"\
-    "$(assertHaveBeenCalled ps)"
+    "$(assert_have_been_called ps)"
 }
 
 
@@ -52,6 +52,6 @@ function test_successful_spy_called_times() {
   ps
   ps
 
-  assertHaveBeenCalledTimes 2 ps
+  assert_have_been_called_times 2 ps
 }